Bugfix: content is not updated on language change
I'm using a dom-if template to switch the content based on the selected
language. It was bound to a function result instead of a computed property,
so it was only computed once (at page/component load).
Now `isPt` is a true computed property and must be treated as such. It will
automatically make all my dom-if templates re-evaluate when the language is
changed by the user.