fix: Use smaller margin for paper-material on smaller screens
On a phone, we're wasting a lot of space in margins instead of occupying it
with actual content (because .paper-material is used as a container in every
page). This changes cuts the margin from 24px to 5px in small screens.
Ideally it would use `min(x%, 24px);` but that is not standard CSS3 and
we cannot rely on browsers to support it.